Coded 2 classes for game prop detection:
FirstVisionProcessor/Opmode - gets avg saturation for each ROI (region of interest) and compares to find greatest, which is most likely game prop

skystoneBasedDetector - this is based off of the skystone detection code (just like 11109's last year) and filters for colors red or blue through YCrCb. Unfortunately red is not being picked up so far so I'll look through that later and possibly use the other class's saturation code if we are not looking for blue.

These classes are different from the python code that I posted in the discord mainly because it looks in specific mark ROIs for significant red/blue fill rather than HSV masking to color object detect the game prop and determine the mark through x-value.
-Ambrose :D